refactor(scripting): rename runScript to scriptUtils
